refactor mtpserver and socks5 tracking state
